% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CDKN1A}
\alias{CDKN1A}
\title{DNA sequence of the CDKN1A promoter}
\format{A character vector containing the 20k promoter sequence of CDKN1A}
\usage{
data(CDKN1A)
}
\description{
A character containing the promoter sequence of CDKN1A, -10 +10 kb from TSS, to be used as an exemplary sequence to test p53retriever functionalities
}
\keyword{datasets}

