% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p53pie.R
\name{p53pie}
\alias{p53pie}
\title{Plot a pie chart of the grades assigned to responsive elements on a DNA sequence}
\usage{
p53pie(p53.table, start.at, stop.at)
}
\arguments{
\item{p53.table}{A dataframe containing the responsive elements, as the one returned from the p53track function.}

\item{start.at}{A numeric value indicating the start coordinate of the site in order to be considered}

\item{stop.at}{A numeric value indicating the stop coordinate start coordinate of the site in order to be considered}
}
\value{
an object of class \code{ggplot}
}
\description{
Plot a pie chart of the grades assigned to responsive elements on a DNA sequence
}
\examples{
data(RE_collection)
p53pie(RE_collection,-10000,10000)
}

