% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p53retriever.R
\docType{package}
\name{p53retriever-package}
\alias{p53retriever}
\alias{p53retriever-package}
\title{detection and display of p53 putative response elements (REs) on DNA sequences}
\description{
p53retriever implements a pattern search algorithm that maps p53
response elements (REs) and ranks them according to predicted transactivation
potentials in five classes. Besides canonical, full site REs, p53tracker
contains specific pattern searches for non-canonical half sites and 3/4 sites.
REs can be displayed along the input sequence in a graphical layout.
}

