% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p53s3q.R
\name{p53s3q}
\alias{p53s3q}
\title{Locate candidate p53 responsive elements (3/4 sites) on a DNA sequence}
\usage{
p53s3q(seq.ini)
}
\arguments{
\item{seq}{A character string containing the sequence. The sequence must be composed exclusively of DNA bases (a,c,t,g)}
}
\value{
A dataframe containing the responsive elements located on the input sequence.
}
\description{
Locate candidate p53 responsive elements (3/4 sites) on a DNA sequence
}

