% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p53sf.R
\name{p53sf}
\alias{p53sf}
\title{Locate candidate p53 responsive elements (full and half) on a DNA sequence}
\usage{
p53sf(seq.ini)
}
\arguments{
\item{seq.ini}{A character string containing the sequence. The sequence must be composed exclusively of DNA bases (a,c,t,g)}
}
\value{
A dataframe containing the responsive elements located on the input sequence.
}
\description{
Locate candidate p53 responsive elements (full and half) on a DNA sequence
}

