% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pwm.1}
\alias{pwm.1}
\title{pwm.1 model matrix}
\format{An object of class \code{matrix}, with 4 rows and 10 columns}
\usage{
data(bkg.models)
}
\description{
pwm.1 is the half site score matrix used by p53retriever to select an initial set of RE full site candidates
}
\keyword{datasets}

