% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pwm.3}
\alias{pwm.3}
\title{pwm.3 model matrix}
\format{An object of class \code{matrix}, with 4 rows and 15 columns}
\usage{
data(bkg.models)
}
\description{
pwm.3 is the score matrix used by p53retriever to select an initial set of RE 3/4 site candidates (3Q234)
}
\keyword{datasets}

