% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pwm.5}
\alias{pwm.5}
\title{pwm.5 model matrix}
\format{An object of class \code{matrix}, with 4 rows and 20 columns}
\usage{
data(bkg.models)
}
\description{
pwm.5 is the score matrix used by p53retriever to select an initial set of RE 3/4 site candidates (3Q124)
}
\keyword{datasets}

