% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pwm.6}
\alias{pwm.6}
\title{pwm.6 model matrix}
\format{An object of class \code{matrix}, with 4 rows and 10 columns}
\usage{
data(bkg.models)
}
\description{
pwm.6 is the half site score matrix used by p53retriever to select an initial set of RE half site candidates
}
\keyword{datasets}

