% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RE_collection}
\alias{RE_collection}
\title{p53 retriever results on a set of human promoters}
\format{A data frame with 3062 rows and 17 variables:
\itemize{
\item HGNC: gene symbol (e.g. TP73)
\item ID: sequence ID (e.g. uc001akp.3)
\item chr: chromosome (e.g. chr1)
\item chr: strand (e.g. +)
\item hg19.TSS: hg19 genomic coordinate of the TSS (e.g. 3569128)
\item hg19.start: hg19 genomic coordinate of the RE's start (e.g. 3559376)
\item hg19.stop: hg19 genomic coordinate of the RE's stop (e.g. 3559405)
\item start: relative coordinate of the RE's start with respect to the TSS (e.g. -9752)
\item stop: relative coordinate of the RE's stop with respect to the TSS (e.g. -9723)
\item spacer: spacer length of the RE (e.g. 10)
\item WW1: sequence of WW1 in the RE (e.g. CT)
\item WW2: sequence of WW2 in the RE (e.g. AT)
\item n.mismatch: number of mismatches with respect to the canonical p53 RE (e.g. 2)
\item mismatch.string: string encoding the position of mismatches in the RE with respect to the canonical one (0 = match, 1 = mismatch, n= spacer, e.g. 0000100000nnnnnnnnnn0000000010)
\item sequence: RE sequence (including the spacer, e.g. gaacctgcttcatttaaatggagcatgtgt)
\item label: mismatch label assigned to the RE by p53retriever, according to the number, the nature and the position of mismatches (0 = no mismatches, A = mismatch in C or G positions, B = mismatch in W positions, C = mismatch in R or Y "external" positions, D = mismatch in R or Y "internal" positions, 3Q = three-quarter site, half = half site)
\item grade: functional score assigned to the RE by p53retriever, according to the number, the nature and the position of mismatches (5 = high, 4 = moderate, 3 = slight, 2 = poor, 1= unlikely)
}}
\usage{
data(RE_collection)
}
\description{
A dataset containing all p53 REs found by p53retreiever on a set of  253 human promoters, to be used as an exemplary dataframe to test p53retriever functionalities
}
\keyword{datasets}

